Move db_lock to utils module
