+++ api/services/users.py
@@ -7,7 +7,7 @@ from passlib.context import CryptContext
 
 from api.models import User
 from api.orm_bootstrap import SessionLocal
-from api.app_state import db_lock
+from api.utils.db_lock import db_lock
 
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
