+++ api/utils/db_lock.py
@@ -1,0 +1,3 @@
+import threading
+
+db_lock = threading.RLock()
